still enjoying my project

+++ app/api/posts/[postId]/route.ts
@@ -1,5 +1,7 @@
 import { NextResponse } from "next/server";
 import prisma from "@/lib/prismadb";
+import { getServerSession } from "next-auth";
+import { authOptions } from "@/app/api/auth/[...nextauth]/route";
 
 // !! GET
 export async function GET(
@@ -11,6 +13,13 @@ export async function GET(
       where: {
         id: params.postId,
       },
+      include: {
+        author: {
+          select: {
+            name: true,
+          },
+        },
+      },
     });
 
     return NextResponse.json(post);
@@ -20,11 +29,17 @@ export async function GET(
   }
 }
 
-// !! PATCH
-export async function PATCH(
+// !! PUT
+export async function PUT(
   req: Request,
   { params }: { params: { postId: string } }
 ) {
+  const session = await getServerSession(authOptions);
+
+  if (!session) {
+    return NextResponse.json({ error: "Not authenticated" }, { status: 401 });
+  }
+
   const { title, content, links, selectedCategory, imageUrl, publicId } =
     await req.json();
 
@@ -55,6 +70,12 @@ export async function DELETE(
   req: Request,
   { params }: { params: { postId: string } }
 ) {
+  const session = await getServerSession(authOptions);
+
+  if (!session) {
+    return NextResponse.json({ error: "Not authenticated" }, { status: 401 });
+  }
+
   try {
     const post = await prisma.post.delete({
       where: {
